requirements.txt
fastapi
uvicorn
python-whois
dnspython
requests
